@@ -4,63 +4,59 @@ description: mbm_elearning, A notes sharing platform.
 version: 7.1.3+20
 
 environment:
-  sdk: ">=2.15.1 <3.0.0"
+  sdk: ">=3.2.1 <4.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-
-  cupertino_icons: ^1.0.3
-  uuid: ^3.0.4
-  http: ^0.13.3
-  firebase_core: ^1.20.0
-  firebase_analytics: ^9.3.0
-  url_launcher: ^6.0.9
-  flutter_bloc: ^7.3.1
-  equatable: ^2.0.3
-  firebase_auth: ^3.6.1
-  google_sign_in: ^5.4.0 # need to setup with double condition
-  share_plus: ^4.0.10
-  carousel_slider: ^4.0.0
-  firebase_messaging: ^12.0.1
-  flutter_linkify: ^5.0.2
-  introduction_screen: ^3.0.2
-  sqflite: ^2.0.0+4 # android ios
-  connectivity_plus: ^2.3.6
-  cloud_firestore: ^3.4.2
-  shared_preferences: ^2.0.15
-  firebase_dynamic_links: ^4.3.3
-  flutter_svg: 
-  permission_handler: ^10.0.0 # android and ios // need to setup with double condition
-  provider: ^6.0.3
-  scrollable_positioned_list: ^0.3.2
+# flutter pub upgrade --major-versions
+  cupertino_icons: ^1.0.6
+  uuid: ^4.2.2
+  http: ^1.1.2
+  firebase_core: ^2.24.2
+  firebase_analytics: ^10.7.4
+  url_launcher: ^6.2.2
+  flutter_bloc: ^8.1.3
+  equatable: ^2.0.5
+  firebase_auth: ^4.15.3
+  google_sign_in: ^6.2.1 # need to setup with double condition
+  share_plus: ^7.2.1
+  carousel_slider: ^4.2.1
+  firebase_messaging: ^14.7.9
+  flutter_linkify: ^6.0.0
+  introduction_screen: ^3.1.12
+  sqflite: ^2.3.0 # android ios
+  connectivity_plus: ^5.0.2
+  cloud_firestore: ^4.13.6
+  shared_preferences: ^2.2.2
+  firebase_dynamic_links: ^5.4.8
+  flutter_svg: 2.0.9
+  permission_handler: ^11.1.0 # android and ios // need to setup with double condition
+  provider: ^6.1.1
+  scrollable_positioned_list: ^0.3.8
   linkfy_text: ^1.1.5
-  googleapis: ^9.1.0
-  file_picker: ^4.3.3
-  googleapis_auth: ^1.3.1
-  html: ^0.15.0
+  googleapis: ^12.0.0
+  file_picker: ^5.2.0+1
+  googleapis_auth: ^1.4.1
+  html: ^0.15.4
   flutter_html: ^3.0.0-alpha.5
   flutter_html_all: ^3.0.0-alpha.5
   dropdown_search: ^5.0.2
-  html_editor_enhanced: ^2.5.0 # warning flutter build web --web-renderer html
-  awesome_notifications: ^0.7.0-beta.5 #android ios
+  html_editor_enhanced: ^2.5.1 # warning flutter build web --web-renderer html
+  awesome_notifications: 0.7.4+1 #android ios
   date_time_picker: ^2.1.0
-  firebase_in_app_messaging: ^0.6.0+20 #android ios
-  tutorial_coach_mark: ^1.2.3
-  calendar_timeline: ^1.0.5
-  chips_choice_null_safety: ^2.0.6
+  firebase_in_app_messaging: ^0.7.4+8 #android ios
+  tutorial_coach_mark: ^1.2.11
   lottie: any
-  update_available: ^2.2.0-dev.1 # android and ios
-  path_provider: ^2.0.11
-  webview_flutter: '>=2.0.4 <4.0.0'
+  update_available: ^3.0.0 # android and ios
+  path_provider: ^2.1.1
+  webview_flutter: ^4.4.2
   
 
 
 dev_dependencies:
-  flutter_lints: ^2.0.1
-  flutter_flavorizr: ^2.1.3
-  flutter_test:
-    sdk: flutter
+  flutter_lints: ^3.0.1
+  flutter_flavorizr: ^2.2.1
 
 flutter:
 
