@@ -10,8 +10,8 @@ import 'package:shared_preferences/shared_preferences.dart';
 
 class SocialSigninButton extends StatelessWidget {
   const SocialSigninButton({
-    Key? key,
-  }) : super(key: key);
+    super.key,
+  });
 
   @override
   Widget build(BuildContext context) {
