@@ -24,7 +24,7 @@ class ModalProgressHUD extends StatelessWidget {
   final Widget child;
 
   const ModalProgressHUD({
-    Key? key,
+    super.key,
     required this.inAsyncCall,
     this.opacity = 0.3,
     this.color = Colors.grey,
@@ -32,7 +32,7 @@ class ModalProgressHUD extends StatelessWidget {
     this.offset,
     this.dismissible = false,
     required this.child,
-  }) : super(key: key);
+  });
 
   @override
   Widget build(BuildContext context) {
